Add minimal testing support
Short of sending actual Http requests to a localhost:port combination
served by a web server, it was previously hard to test routing/controllers.
These changes hopefully make it slightly easier. They add
the ability to examine the status_code and data of the last HTTP
request. With these changes, it is possible to mock sending,
and receiving the results of, a HTTP request.

* Support examining the status_code and data of the last HTTP request.