Fix the method for including controller in router
Previously, the router used `include` to evaluate the controller receiving
the method. Calling `include` worked perfectly on the server. Because
Apache started a new PHP environment for each request, `include` could
not be called twice on the same class file. However, running tests
with PHPunit maintains the same PHP environment for
the entire test suite. Routing to a controller more than once caused
reevaluation of the class, resulting in a "Cannot redeclare class"
error. Using `include_once` to initially evaluate the controller file
ensured PHP never attempted to redeclare the class, and fixed the bug.