Update request variables parsing
Update the parsing of variables with `Controller#read_requests`. Add an
option, turned off by default, will not add a variable to the
key/value store of keys and request variables if the variable is null.
This feature is very useful for when you want to update a model instance,
but can not guarantee all fields will be included in the HTTP request.

* Update the parsing of variables with `Controller#read_requests`.